Split out common insights
