+++ tests/unit/test_large_video_file_insight.py
@@ -1,6 +1,6 @@
 from unittest.mock import Mock
 
-from launchpad.size.insights.common import LargeVideoFileInsight
+from launchpad.size.insights.common.large_videos import LargeVideoFileInsight
 from launchpad.size.insights.insight import InsightsInput
 from launchpad.size.models.common import BaseAppInfo, FileAnalysis, FileInfo
 from launchpad.size.models.insights import LargeVideoFileInsightResult
